refactor: move EmptyView component to cards directory
